Emulate TrackingVH using WeakVH (#6662)
This PR pulls the upstream change, Emulate TrackingVH using WeakVH
(https://github.com/llvm/llvm-project/commit/8a6238201f015729a47691c62808a23ec8525096),
into DXC.

Here's the summary of the change:

> This frees up one slot in the HandleBaseKind enum, which I will use
later to add a new kind of value handle. The size of the HandleBaseKind
enum is important because we store a HandleBaseKind in
>   the low two bits of a (in the worst case) 4 byte aligned pointer.
> 
>   Reviewers: davide, chandlerc
> 
>   Subscribers: mcrosier, llvm-commits
> 
>   Differential Revision: https://reviews.llvm.org/D32634

This is part 2 of the fix for #6659.