[SPIRV] Register all decls for a variable. (#6969)
Some variable can have multiple decls. In one case, there could be a
declaration of a const static member variable that is later defined. All
of these decls need to be in astDecls and associated with the same
variable.

The current code will only add the decl for the defintion. This leads to
problems when trying to find the variable for the declaration.

Fixes #6787